requirements.txt
openai
langchain
langchain-community
chromadb
streamlit
fastapi
uvicorn
redis
python-dotenv
pypdf